requirements.txt
django
faker
djangorestframework
django-grappelli

